+++ tests/dynamic_spec.rs
@@ -2,11 +2,11 @@
 
 use core::mem;
 
-use cell_grid::dynamic::{Grid, IncompatibleRowSize};
+use cell_grid::{DynamicGrid, IncompatibleRowSize};
 
 #[test]
 fn can_push_rows() {
-    let mut grid: Grid<(i32, i32)> = Grid::new();
+    let mut grid: DynamicGrid<(i32, i32)> = DynamicGrid::new();
     for y in 0..5 {
         grid.push_row((0..5).map(|x| (x, y))).unwrap();
     }
@@ -15,7 +15,7 @@ fn can_push_rows() {
 
 #[test]
 fn cannot_push_row_of_incompatible_width() {
-    let mut grid: Grid<i32> = Grid::new();
+    let mut grid: DynamicGrid<i32> = DynamicGrid::new();
     grid.push_row(0..10).unwrap();
     let _: IncompatibleRowSize = grid.push_row(0..11).unwrap_err();
     let _: IncompatibleRowSize = grid.push_row(0..9).unwrap_err();
@@ -24,25 +24,25 @@ fn cannot_push_row_of_incompatible_width() {
 
 #[test]
 fn can_create_from_size_and_init_function() {
-    let grid = Grid::new_with(5, 5, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(5, 5, |x, y| (x, y));
     assert_eq!(grid.get(2, 3), Some(&(2, 3)));
 }
 
 #[test]
 fn get_return_null_when_x_is_out_of_bounds() {
-    let grid = Grid::new_with(5, 5, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(5, 5, |x, y| (x, y));
     assert_eq!(grid.get(5, 0), None);
 }
 
 #[test]
 fn can_create_from_size_and_default() {
-    let grid = Grid::<bool>::new_with_default(5, 5);
+    let grid = DynamicGrid::<bool>::new_with_default(5, 5);
     assert_eq!(grid.get(2, 3), Some(&false));
 }
 
 #[test]
 fn can_create_from_row_major_iter() {
-    let grid = Grid::new_from_iter(2, [1, 2, 3, 4]).unwrap();
+    let grid = DynamicGrid::new_from_iter(2, [1, 2, 3, 4]).unwrap();
     assert_eq!(grid.get(0, 0), Some(&1));
     assert_eq!(grid.get(1, 0), Some(&2));
     assert_eq!(grid.get(0, 1), Some(&3));
@@ -51,23 +51,23 @@ fn can_create_from_row_major_iter() {
 
 #[test]
 fn cannot_create_from_invalid_row_major_iter() {
-    let _: IncompatibleRowSize = Grid::new_from_iter(2, [1, 2, 3, 4, 5]).unwrap_err();
+    let _: IncompatibleRowSize = DynamicGrid::new_from_iter(2, [1, 2, 3, 4, 5]).unwrap_err();
 }
 
 #[test]
 fn cannot_create_from_invalid_row_major_iter_wth_invalid_width() {
-    let _: IncompatibleRowSize = Grid::new_from_iter(0, [1]).unwrap_err();
+    let _: IncompatibleRowSize = DynamicGrid::new_from_iter(0, [1]).unwrap_err();
 }
 
 #[test]
 fn can_create_empty_grid_from_iter() {
-    let grid = Grid::<i32>::new_from_iter(0, []).unwrap();
+    let grid = DynamicGrid::<i32>::new_from_iter(0, []).unwrap();
     assert!(grid.is_empty());
 }
 
 #[test]
 fn can_mutate_cell() {
-    let mut grid = Grid::new_with_default(2, 2);
+    let mut grid = DynamicGrid::new_with_default(2, 2);
     assert_eq!(grid.get(1, 0), Some(&0));
     *grid.get_mut(1, 0).unwrap() = 2;
     assert_eq!(grid.get(1, 0), Some(&2));
@@ -77,14 +77,14 @@ fn can_mutate_cell() {
 
 #[test]
 fn can_iterate_cells() {
-    let grid = Grid::new_with(2, 2, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(2, 2, |x, y| (x, y));
     let rows: Vec<&(usize, usize)> = grid.cells().collect();
     assert_eq!(rows, [&(0, 0), &(1, 0), &(0, 1), &(1, 1)]);
 }
 
 #[test]
 fn can_iterate_cell_mutably() {
-    let mut grid = Grid::new_with(3, 3, |x, y| (x, y));
+    let mut grid = DynamicGrid::new_with(3, 3, |x, y| (x, y));
     for (x, y) in grid.cells_mut() {
         mem::swap(x, y);
     }
@@ -93,14 +93,14 @@ fn can_iterate_cell_mutably() {
 
 #[test]
 fn can_iterate_rows() {
-    let grid = Grid::new_with(2, 2, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(2, 2, |x, y| (x, y));
     let rows: Vec<&[(usize, usize)]> = grid.rows().collect();
     assert_eq!(rows, [&[(0, 0), (1, 0)], &[(0, 1), (1, 1)],]);
 }
 
 #[test]
 fn can_iterate_cells_overlaping_a_rectangle() {
-    let grid = Grid::new_with(10, 10, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(10, 10, |x, y| (x, y));
     let cells: Vec<&(usize, usize)> = grid.cells_in_rect(4, 5, 2, 3).collect();
     assert_eq!(
         cells,
@@ -110,14 +110,14 @@ fn can_iterate_cells_overlaping_a_rectangle() {
 
 #[test]
 fn out_of_bounds_are_ignored_when_iterating_overlaping_rect() {
-    let grid = Grid::new_with(10, 10, |x, y| (x, y));
+    let grid = DynamicGrid::new_with(10, 10, |x, y| (x, y));
     let cells: Vec<&(usize, usize)> = grid.cells_in_rect(9, 9, 5, 5).collect();
     assert_eq!(cells, [&(9, 9)]);
 }
 
 #[test]
 fn grid_should_be_thread_safe() {
-    assert_thread_safe::<Grid<i32>>();
+    assert_thread_safe::<DynamicGrid<i32>>();
 }
 
 #[test]