+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 
+## [0.1.4] - 2024-12-08
+
+
 ### Added
 
 * Rename `dynamic::Grid` to `DynamicGrid` (the former remains as an alias)
@@ -77,7 +80,8 @@ This is essentially a second attempt at designing the `Grid` API, with two main 
 * `no_std` support (by disabling the `std` feature)
 * `aline-v01` feature for compatiblity with the [`aline`](https://github.com/jcornaz/aline) crate
 
-[Unreleased]: https://github.com/jcornaz/cell-grid/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/jcornaz/cell-grid/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/jcornaz/cell-grid/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/jcornaz/cell-grid/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/jcornaz/cell-grid/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/jcornaz/cell-grid/compare/v0.1.0...v0.1.1